pyproject.toml
[tool.poetry]
name = "bot-front-new"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "tg_bot"}]

[tool.poetry.dependencies]
python = "^3.11"
python-telegram-bot = "^20.3"
aiohttp = "^3.8.4"
python-dotenv = "^1.0.0"
black = "^23.3.0"
isort = "^5.12.0"
flake8 = "^6.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "bot_backend.settings"
filterwarnings = [
    "ignore::UserWarning",
    "ignore::DeprecationWarning"
]


[tool.black]
line-length = 79
target_version = ["py310"]

[tool.isort]
profile = "black"
py_version = 310
multi_line_output = 3

